Update drg_ float to decimal

+++ database/migrations/2020_08_13_195927_create_drg_sample_to_isps_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->id('id');
             $table->integer('number')->nullable();
             $table->date('document_date')->nullable();
-            $table->float('envelope_weight', 8, 2);
+            $table->decimal('envelope_weight', 8, 2);
             $table->text('observation')->nullable();
             $table->foreignId('reception_id')->constrained('drg_receptions');
             $table->foreignId('user_id')->constrained('users');
